test(vitest): split integration spec into parallel CI jobs via SPEC matrix
Split vitest.spec.js (3 366 lines) into three files:

- vitest.core.spec.js        — pool, propagation, error tags, flaky retries, EFD
- vitest.test-management.spec.js — known tests without EFD, test management
- vitest.advanced.spec.js    — libraries capabilities, impacted tests,
                               programmatic API, coverage upload, final status

Add a spec matrix to integration-vitest and pass SPEC to the coverage
script, mirroring the same pattern used for Jest and Cypress.
